gps/requirements.txt
geopy==2.3.0
gpsd_py3==0.3.0
paho_mqtt==1.6.1
